requirements.txt
# No external dependencies required
# This tool uses only Python standard library modules
azure-storage-blob==12.26.0